+++ src/main.cpp
@@ -84,7 +84,7 @@ RED4EXT_C_EXPORT void RED4EXT_CALL Query(PluginInfo* aInfo)
 {
     aInfo->name = L"New Game+";
     aInfo->author = L"not_alphanine";
-    aInfo->version = RED4EXT_SEMVER_EX(build::c_version.major, build::c_version.minor, build::c_version.patch, RED4EXT_V0_SEMVER_PRERELEASE_TYPE_NONE, 0); // Set your version here.
+    aInfo->version = RED4EXT_SEMVER_EX(static_cast<std::uint8_t>(build::c_version.major), static_cast<std::uint8_t>(build::c_version.minor), static_cast<std::uint8_t>(build::c_version.patch), RED4EXT_V0_SEMVER_PRERELEASE_TYPE_NONE, 0); // Set your version here.
     aInfo->runtime = RED4EXT_RUNTIME_INDEPENDENT;
     aInfo->sdk = RED4EXT_SDK_LATEST;
 }
